move setGroupBookmarks into api/common
